Add tests that cover all cases documented in the help.
I did not write tests for all filters, some of them were so simple there were no examples in the documentation.

This includes some failing test cases (commented out) in the following filters:

- link
- merge
- split
- table
- wikilink

I was also unable to test map and template because they import debug which imports some polyfill thing that complains about not running in a browser. Can probably fix this through some kind of mocking or injection.